pyproject.toml
[project]
name = "tavily-serach-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.107.0",
    "openai-agents>=0.2.11",
    "python-decouple>=3.8",
    "tavily-python>=0.7.11",
]
